Linked language and pain service with data service
Now using the real queryString from the running services (ports 5001 and
5003).
If something fails from the service we fallback to direct query
appending to the query string.

Also fixed pain.py service bug - comments range were seperated with ':'
and should be '..'

Also fixed a bug in Clarity UI - calling detect changes when the data is
retrieved.